More instance creation reorg
