+++ src/pykeen/training/slcwa.py
@@ -1,17 +1,18 @@
 """Training KGE models based on the sLCWA."""
 
 import logging
-from typing import Optional
+import warnings
+from typing import Any, Optional
 
 from class_resolver import HintOrType, OptionalKwargs
-from torch.utils.data import DataLoader
+from torch.utils.data import DataLoader, Dataset
 
 from .training_loop import TrainingLoop
 from ..losses import Loss
 from ..models.base import Model
 from ..sampling import NegativeSampler
 from ..triples import CoreTriplesFactory
-from ..triples.instances import SLCWABatch, SLCWASampleType
+from ..triples.instances import BatchedSLCWAInstances, SLCWABatch, SLCWASampleType, SubGraphSLCWAInstances
 from ..typing import FloatTensor, InductiveMode
 
 __all__ = [
@@ -51,7 +52,8 @@ class SLCWATrainingLoop(TrainingLoop[SLCWASampleType, SLCWABatch]):
     ) -> DataLoader[SLCWABatch]:  # noqa: D102
         assert "batch_sampler" not in kwargs
         return DataLoader(
-            dataset=triples_factory.create_slcwa_instances(
+            dataset=create_slcwa_instances(
+                triples_factory,
                 batch_size=batch_size,
                 shuffle=kwargs.pop("shuffle", True),
                 drop_last=drop_last,
@@ -154,3 +156,24 @@ class SLCWATrainingLoop(TrainingLoop[SLCWASampleType, SLCWABatch]):
             report = "This model doesn't support sub-batching and slicing is not possible for sLCWA"
         logger.warning(report)
         raise MemoryError("The current model can't be trained on this hardware with these parameters.")
+
+
+def create_slcwa_instances(
+    triples_factory: CoreTriplesFactory,
+    *,
+    sampler: Optional[str] = None,
+    **kwargs: Any,
+) -> Dataset:
+    """Create sLCWA instances for this factory's triples."""
+    cls = BatchedSLCWAInstances if sampler is None else SubGraphSLCWAInstances
+    if "shuffle" in kwargs:
+        if kwargs.pop("shuffle"):
+            warnings.warn("Training instances are always shuffled.", DeprecationWarning, stacklevel=2)
+        else:
+            raise AssertionError("If shuffle is provided, it must be True.")
+    return cls(
+        mapped_triples=triples_factory._add_inverse_triples_if_necessary(mapped_triples=triples_factory.mapped_triples),
+        num_entities=triples_factory.num_entities,
+        num_relations=triples_factory.num_relations,
+        **kwargs,
+    )
